Extract speech recognition to the separate class
